Hey look, if I do my function pointer casts like this: (void(*)()) (size_t) some_void_pointer then -pedantic doesn't complain! Adding -pedantic back to the -Wall test case.

+++ cplof/src/impl/ccall.c
@@ -17,7 +17,8 @@ label(interp_psl_ccall);
         struct PlofObject *otmp;
 
         cif = (ffi_cif_plus *) ASPTR(a);
-        func = (void(*)()) ASPTR(b);
+        /* this stupid cast is to compile with -pedantic */
+        func = (void(*)()) (size_t) ASPTR(b);
         ad = ARRAY(c);
 
         /* now start preparing the args and return */
